updateUserPollsParticipated() now use long for pollID

+++ src/main/java/Projects/ProjectB/Controller.java
@@ -184,7 +184,7 @@ public class Controller {
 
 	@PutMapping("/users/participatePoll/{userName}")
 	public String updateUserPollsParticipated(@PathVariable String userName, @RequestBody Map<String, String> json) {
-		int pollID = Integer.parseInt(json.get("pollID"));
+		long pollID = Long.parseLong(json.get("pollID"));
 		String creatorUserName = "" + json.get("creator");
 		User creator = userRepository.findByUserName(creatorUserName);
